refactor: MainPage 위치 변경
