+++ src/App.tsx
@@ -3,7 +3,7 @@ import "./App.css";
 import "./PostContent.css";
 import { PATH } from './constants/Path';
 import AdminLoginCallbackPage from './domain/auth/page/AdminLoginCallbackPage';
-import MainPage from './page/MainPage';
+import MainPage from './domain/common/page/MainPage';
 import { Pages } from './constants/Pages';
 import LoginPage from './domain/auth/page/LoginPage';
 import GitHubCallbackPage from './domain/auth/page/GitHubCallbackPage';
